feat: admin and users permissions control for update and create controllers

+++ api/controllers/pets.controller.js
@@ -2,7 +2,8 @@ const createError = require("http-errors");
 const Pet = require("../models/pet.model");
 
 module.exports.list = (req, res, next) => {
-  Pet.find()
+  const { userId } = req.params
+  Pet.find({ owner: userId })
     .populate("owner")
     .then((pets) => res.json(pets))
     .catch(next);
@@ -11,19 +12,21 @@ module.exports.list = (req, res, next) => {
 module.exports.get = (req, res, next) => res.json(req.foundPet);
 
 module.exports.create = (req, res, next) => {
-  Pet.create({ ...req.body, user: req.user })
-    .then((pet) => {
-      console.log("USER", req.user);
-      res.status(201).json(pet);
-    })
+  Pet.create({ ...req.body, owner: req.user.id })
+    .then((pet) => res.status(201).json(pet))
     .catch(next);
 };
 
 module.exports.delete = (req, res, next) => {
-  req.foundPet
-    .delete()
-    .then(() => res.status(204).end())
-    .catch(next);
+
+  if (req.user.role === 'admin' || req.user.id === req.foundPet.owner.id) {
+    req.foundPet
+      .delete()
+      .then(() => res.status(204).end())
+      .catch(next);
+  } else {
+    next(createError(403, 'Forbidden permissions'))
+  }
 };
 
 module.exports.update = (req, res, next) => {
@@ -34,10 +37,17 @@ module.exports.update = (req, res, next) => {
       coordinates: location,
     };
   }
+  if (req.file) {
+    req.body.image = req.file.path
+  }
   Object.assign(req.foundPet, req.body);
+  if (req.user.role === 'admin' || req.user.id === req.foundPet.owner.id) {
+    req.foundPet
+      .save()
+      .then((pet) => res.json(pet))
+      .catch(next);
+  } else {
+    next(createError(403, 'Forbidden permissions'))
+  }
 
-  req.foundPet
-    .save()
-    .then((pet) => res.json(pet))
-    .catch(next);
 };
